move ExternalLinkFlag back into controller because of instance dependency...
