--- app/scripts/components/common/card/external-link-flag.tsx
@@ -1,39 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import { multiply, themeVal } from '@devseed-ui/theme-provider';
-import { CollecticonExpandTopRight } from '@devseed-ui/collecticons';
-import { variableGlsp } from '$styles/variable-utils';
-
-// @TODO: This should be moved elsewhere
-
-const ExternalLinkMark = styled.div`
-  display: flex;
-  align-items: center;
-  position: absolute;
-  top: ${variableGlsp(0.25)};
-  right: ${variableGlsp(0.25)};
-  padding: ${variableGlsp(0.125)} ${variableGlsp(0.25)};
-  background-color: ${themeVal('color.primary')};
-  color: ${themeVal('color.surface')};
-  text-transform: none;
-  border-radius: calc(
-    ${multiply(themeVal('shape.rounded'), 2)} - ${variableGlsp(0.125)}
-  );
-  z-index: 1;
-`;
-
-const FlagText = styled.div`
-  display: inline;
-  font-weight: bold;
-  font-size: 0.825rem;
-  margin-right: ${variableGlsp(0.25)};
-`;
-
-export function ExternalLinkFlag() {
-  return (
-    <ExternalLinkMark>
-      <FlagText>External Link</FlagText>
-      <CollecticonExpandTopRight size='small' meaningful={false} />
-    </ExternalLinkMark>
-  );
-}
